requirement.txt
asgiref==3.7.2
autocommand==2.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.13
django-elasticsearch-dsl==8.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.5.0
drf-yasg==1.21.7
elastic-transport==8.12.0
elastic2-doc-manager==1.0.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
future==1.0.0
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
openapi-codec==1.3.2
packaging==24.0
pymongo==3.12.3
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
serializers==0.2.4
simplejson==3.19.2
six==1.16.0
sqlparse==0.2.4
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==1.26.18
zipp==3.18.0
